Move domain layer to separate module
