+++ Domain/src/main/java/com/tcs/sample/domain/model/ProductList.kt
@@ -1,4 +1,4 @@
-package com.tcs.sample.cleanmvvm.domain.model
+package com.tcs.sample.domain.model
 
 data class ProductList(
     val products: List<ProductDetail>? = null
