+++ Domain/src/main/java/com/tcs/sample/domain/usecases/GetProductUseCase.kt
@@ -1,0 +1,8 @@
+package com.tcs.sample.cleanmvvm.domain.usecases
+
+import com.tcs.sample.domain.repository.RemoteDataSource
+import javax.inject.Inject
+
+class GetProductUseCase @Inject constructor(private val remoteDataSource: RemoteDataSource) {
+    suspend fun getProduct(id: Int) = remoteDataSource.getProduct(id)
+}
