+++ Domain/src/main/java/com/tcs/sample/domain/usecases/GetProductsListUseCase.kt
@@ -1,6 +1,6 @@
-package com.tcs.sample.cleanmvvm.domain.usecases
+package com.tcs.sample.domain.usecases
 
-import com.tcs.sample.cleanmvvm.domain.repository.RemoteDataSource
+import com.tcs.sample.domain.repository.RemoteDataSource
 import javax.inject.Inject
 
 class GetProductUseCase  @Inject constructor(private val remoteDataSource: RemoteDataSource) {
