--- app/src/main/java/com/tcs/sample/cleanmvvm/domain/usecases/GetProductsListUseCase.kt
@@ -1,8 +1,0 @@
-package com.tcs.sample.cleanmvvm.domain.usecases
-
-import com.tcs.sample.cleanmvvm.domain.repository.RemoteDataSource
-import javax.inject.Inject
-
-class GetProductsListUseCase @Inject constructor(private val remoteDataSource: RemoteDataSource) {
-    suspend fun getProductList() = remoteDataSource.getProductsList()
-}
