+++ app/src/main/java/com/tcs/sample/cleanmvvm/ui/productlist/ProductListViewModel.kt
@@ -3,8 +3,8 @@ package com.tcs.sample.cleanmvvm.ui.productlist
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.tcs.sample.cleanmvvm.domain.model.ProductList
-import com.tcs.sample.cleanmvvm.domain.usecases.GetProductsListUseCase
+import com.tcs.sample.domain.usecases.GetProductsListUseCase
+import com.tcs.sample.domain.model.ProductList
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
