+++ app/src/main/res/layout/layout_product_details_fragment.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-        <variable name="product" type="com.tcs.sample.cleanmvvm.domain.model.ProductDetail" />
+        <variable name="product" type="com.tcs.sample.domain.model.ProductDetail" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
