+++ app/src/main/res/layout/layout_product_item_row.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-        <variable name="product" type="com.tcs.sample.cleanmvvm.domain.model.ProductDetail" />
+        <variable name="product" type="com.tcs.sample.domain.model.ProductDetail" />
     </data>
 
     <androidx.cardview.widget.CardView
