+++ app/src/test/java/com/tcs/sample/cleanmvvm/domain/usecase/GetProductUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.tcs.sample.cleanmvvm.domain.usecase
 
 import com.tcs.sample.cleanmvvm.data.repository.ProductsRepositoryImpl
-import com.tcs.sample.cleanmvvm.domain.model.ProductDetail
+import com.tcs.sample.domain.model.ProductDetail
 import com.tcs.sample.cleanmvvm.domain.usecases.GetProductUseCase
 import io.mockk.coEvery
 import io.mockk.mockk
