+++ app/src/test/java/com/tcs/sample/cleanmvvm/domain/viewmodel/ProductListViewModelTest.kt
@@ -1,8 +1,8 @@
 package com.tcs.sample.cleanmvvm.domain.viewmodel
 
-import com.tcs.sample.cleanmvvm.domain.model.ProductDetail
+import com.tcs.sample.domain.model.ProductDetail
 import com.tcs.sample.cleanmvvm.domain.model.ProductList
-import com.tcs.sample.cleanmvvm.domain.usecases.GetProductsListUseCase
+import com.tcs.sample.domain.usecases.GetProductsListUseCase
 import com.tcs.sample.cleanmvvm.ui.productlist.ProductListViewModel
 import io.mockk.coEvery
 import io.mockk.mockk
